Use PID inspired input filtering
Due to temperature changes while running, the value on the analog in from
the knob would change enough over time to move the rolling average and trip
the update logic. This patch uses a PID controller to slowly move the
assumed value of the pot to match the reading of the analog in. This gives
the added bonus of a Derivative term which is a very good indicator of
whether the knob is moving or not.